% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_createNCDF.R
\name{CMIP5.createNCDF}
\alias{CMIP5.createNCDF}
\title{CMIP5_createNCDF writes annual files for various weather variables according
to JULES accepted format}
\usage{
CMIP5.createNCDF(var, scenario, model, starttime, endtime, hires.map, subdata,
  out_path)
}
\arguments{
\item{...}{}
}
\value{
...
}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed
do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
est laborum.
}
\examples{
# ....
}
\author{
Zed Zulkafli
}

