% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_getIndex.R
\name{CMIP5.getIndex}
\alias{CMIP5.getIndex}
\title{CMIP5_getIndex reads the full time series in climate model output files and
returns the time series in the user specified spatial (averaged) and temporal domain}
\usage{
CMIP5.getIndex(var, CCscenarios, models, startyear, endyear, CMIP5_paths,
  bboxes, out_path)
}
\arguments{
\item{CCscenarios}{is a string vector containing climate scenarios names.}

\item{models}{is a string vector containing climate model names.}

\item{bboxes}{is list of matrices containing the bounding box(es) associated
to all climate indices of interest. The list can be generated using Define.ClimateIndex()}

\item{out_path}{is a string vector containing the user's working directory name}

\item{vars}{is a string vector containing climate variable names.}

\item{starttime}{is a numeric vector of size 1 containing the start of the time period of interest}

\item{endtime}{is a numeric vector of size 1 containing the end of the time period of interest}
}
\value{
NULL. Results are saved in files in the user's working directory.
}
\description{
This function performs multiple tasks in series.
1. The function reads through each of the GCM output files in a list and
   populates the time series based on the calendar type (gregorian, 365,
   360 days, etc., varies from model to model).
2. The function identifies the starting points and endpoints in time
   series within each file that fall within the time domain from the user
   input and extracts the data.
3. The function clips the global time series using the bounding box(es)
   from the user input, then calculates the spatial mean.
4. The function saves the time series of spatial mean in rda files.
}
\examples{
# CMIP5.getIndex(var 		= "tos", CCscenarios	= "historical", models 		= "CSIRO-QCCCE/CSIRO-Mk3-6-0", starttime	= ISOdate(1976,01,01,0,0), endtime		= ISOdate(2005,12,31,0,0), CMIP5_paths = "/var/data/CMIP5", bboxes = cbind(c(210,270,270,210),c(-5,-5,5,5)), out_path = "/home/me/")
}
\author{
Zed Zulkafli
}

