% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_getPaths.R
\name{CMIP5.getpaths}
\alias{CMIP5.getpaths}
\title{CMIP5_getPaths identifies relevant GCM output files on the data directory
based on the user preferences}
\usage{
CMIP5.getpaths(vars, ensemble, models, CCscenarios, data_path)
}
\arguments{
\item{vars}{is a string vector containing climate variable names.}

\item{ensemble}{is a string vector containing model ensemble names.}

\item{models}{is a string vector containing climate model names.}

\item{CCscenarios}{is a string vector containing climate scenarios names.}

\item{data_path}{is a string vector containing the CMIP5 data directory name}
}
\value{
CMIP5_paths is a 3-tiered list containing climate model filenames,
arranged by climate variables, climate scenarios, and climate model names
}
\description{
This function scans the CMIP5 data directory and returns the
names of GCM output files that contain the variables, scenarios, ensembles,
and models of the user's interest
}
\examples{
# CMIP5_paths <- CMIP5.getpaths(	vars		= "tos", ensemble	= "r1i1p1", models  	= "CSIRO-QCCCE/CSIRO-Mk3-6-0", CCscenarios	= "historical", data_path	= "/var/data/CMIP5/")
}
\author{
Zed Zulkafli
}

