% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_plotIndex.R
\name{CMIP5.plotIndex}
\alias{CMIP5.plotIndex}
\title{CMIP5_getPaths identifies relevant GCM output files on the data directory
based on the user preferences}
\usage{
CMIP5.plotIndex(climateIndices, models, CCscenarios, out_path)
}
\arguments{
\item{models}{is a string vector containing climate model names.}

\item{CCscenarios}{is a string vector containing climate scenarios names.}

\item{out_path}{is a string vector containing the user's working directory name}

\item{ClimateIndices}{is a string vector containing climate variable names}

\item{out_path}{}
}
\value{
NULL. Plots are saved in the
}
\description{
This function scans the CMIP5 data directory and returns the
names of GCM output files that contain the variables, scenarios, ensembles,
and models of the user's interest
}
\examples{
# CMIP5(climateIndices =climateIndices, models="CSIRO-QCCCE/CSIRO-Mk3-6-0", CCscenarios="historical", out_path = "/home/me")
}
\author{
Zed Zulkafli
}

