% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_regridTS.R
\name{CMIP5.regridTS}
\alias{CMIP5.regridTS}
\title{CMIP5_regridTS reads the full time series in climate model output files and
returns the time series in the user specified spatial (interpolated to the user
grid) and temporal domain}
\usage{
CMIP5.regridTS(var, CCscenarios, models, startyear, endyear, CMIP5.paths,
  ModelDomain, out.path)
}
\arguments{
\item{...}{}
}
\value{
...
}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed
do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
est laborum.
}
\examples{
# ....
}
\author{
Zed Zulkafli
}

