% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_defineClimateIndex.R
\name{Define.ClimateIndex}
\alias{Define.ClimateIndex}
\title{CMIP5_defineClimateIndex reads in user input of Climate Indices and returns
their corresponding geographical bounding boxes}
\usage{
Define.ClimateIndex(climateIndices)
}
\arguments{
\item{climateIndices}{is a string vector of Climate Indices names}
}
\value{
bboxes is a list of bounding box definions in matrix format
}
\description{
This function reads in user input of Climate Indices names,
and returns their corresponding geographical bounding boxes for subsequent
data extraction and analysis (e.g spatial averaging). The bounding box
definitions are currently limited to ENSO and IOD indices.
references: http://www.esrk.noaa.gov/psd/data/climateindices/list/
           http://www.bom.gov.au/climate/IO/about_IOD.shtml
}
\examples{
# bboxes  <- Define.ClimateIndex(  c("NINO_3", "NINO_4", "NINO_3.4", "NINO_1+2", "IOD_west", "IOD_east")
}
\author{
Zed Zulkafli
}

