% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMIP5_getDomain.R
\name{Define.Domain}
\alias{Define.Domain}
\title{CMIP5 getDomain defines a distributed model grid at the spatial limits and resolution provided by the user}
\usage{
Define.Domain(lat_range, lon_range, res)
}
\arguments{
\item{lat_range}{is a numeric vector of size 2 containing the lower and higher latitudinal limits
(acceptable values -90 to +90)}

\item{lon_range}{is a numeric vector of size 2 containing the left and right longitudinal limits
(acceptable values -180 to +180)}

\item{res}{is a numeric vector of size 1 containing the required distance between points
(uniform in x and y direction)}
}
\value{
Model_Domain is a list containing
			1) bboxes: a 2 x 2 matrix of the bounding box
			2) hires.grid: a spatial grid data frame of points
			within this bounding box, separated distances equal to param res
}
\description{
CMIP5 getDomain defines a distributed model grid at the spatial limits and resolution provided by the user
}
\examples{
# ModelDomain <- Define.Domain(lat_range = c(-11,1), c(-80, -70), res = 0.125)
}
\author{
Zed Zulkafli

This function reads in the coordinate limits and spatial resolution of the analysis/modelling domain
and returns a list containing the 1) bounding box and 2) a spatial grid data frame of points interpolated
within this bounding box
}

