% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NNinterp.R
\name{NNinterp}
\alias{NNinterp}
\title{NNinterp interpolates data on one grid set to a second grid set}
\usage{
NNinterp(map_ts, hires.map)
}
\arguments{
\item{map_ts}{is a list containing [[1]]: a time series data frame [[2]]: a spatial grid data frame associated with [[1]]. map_ts[[2]]$ID is the identifier (can be string or numeric) for each data point in the data frame. Map_ts relates to the source data that requires interpolation}

\item{hires.map}{is a spatial grid data frame that is the target interpolation grid}
}
\value{
RegridTS is a list containing [[1]]: an interpolated time series data frame [[2]]: a spatial grid data frame associated with [[1]].
}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed
do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
est laborum.
}
\examples{
# ....
}
\author{
Zed Zulkafli
}

